reorg and renamed workflow steps, steps category
