+++ cognite/neat/workflows/steps/lib/__init__.py
@@ -1,5 +1,6 @@
-from .cdf_resources import *  # noqa
-from .transformation_rules import *  # noqa
-from .graphs import *  # noqa
-from .graph_data_capture import *  # noqa
-from .data_modeling import *  # noqa
+from cognite.neat.workflows.steps.lib.graph_extractor import *  # noqa
+from cognite.neat.workflows.steps.lib.graph_transformer import *  # noqa
+from cognite.neat.workflows.steps.lib.graph_loader import *  # noqa
+from cognite.neat.workflows.steps.lib.graph_store import *  # noqa
+from cognite.neat.workflows.steps.lib.rules_exporter import *  # noqa
+from cognite.neat.workflows.steps.lib.rules_parser import *  # noqa
