+++ cognite/neat/workflows/steps/step_model.py
@@ -1,9 +1,20 @@
 from abc import ABC, abstractmethod
+from enum import StrEnum
 from typing import ClassVar, TypeVar
 from pydantic import BaseModel, ConfigDict
 from cognite.neat.app.monitoring.metrics import NeatMetricsCollector
 
 from cognite.neat.workflows.model import WorkflowConfigItem, WorkflowConfigs
+
+
+class StepCategory(StrEnum):
+    RulesImporter = "rules | importer"
+    RulesParser = "rules | parser"
+    RulesExporter = "rules | exporter"
+    GraphExtractor = "graph | extractor"
+    GraphTransformer = "graph | transformer"
+    GraphStore = "graph | store"
+    GraphLoader = "graph | loader"
 
 
 class Config(BaseModel):
